==> code/qss20_groupcode/dol_staffing/00_final_basic_cleaning_opm.ipynb <==
import pandas as pd
import os
import itertools
from itertools import product
import numpy as np
import re 

os.getcwd()

# cleaning functions

# subset opm data to correct agency and occupation
def subset_data(df):
    df_dlwh = df[df['AGYSUB'] == "DLWH"].copy()
    df_investigators = df_dlwh[(df_dlwh['OCC'] == 1849)|(df_dlwh["OCC"]=='1849')]
    return df_investigators


# adding year and quarter columns to dataframe
def add_yqs(df, y, q):
    df['YEAR'] = y
    df['QUARTER'] = q
    return df


# meta function to subset data, add columns, and create list of those dfs
def create_df(df, y, q):
    subset = subset_data(add_yqs(df, y, q))    
    return subset

# start of opm cleaning, creating naming patterns
name_pattern = "FACTDATA_{mon}{year}.TXT"
all_years = np.arange(2011, 2021)
all_quarters = ['MAR', 'JUN', 'SEP', 'DEC']
all_combos = list(product(all_years, all_quarters))

all_names = [name_pattern.format(mon = x[1], 
                                 year = x[0]) for x in all_combos]
all_names

#feeding meta function individual dfs and returning list of cleaned dfs
df_list = []

count = 0;
for name in all_names:
    df = pd.read_csv(name)
    year = all_combos[count][0]
    quarter = all_combos[count][1]
    to_append = create_df(df, year, quarter)
    df_list.append(create_df(df, year, quarter))
    count += 1

#concatenating all quarterly/yearly dfs together
full_opm_dataset = pd.concat(df_list)
full_opm_dataset["SALARY"] = [float(re.sub("[\$,]", "", str(full_opm_dataset.SALARY.iloc[i]))) for i in range(len(full_opm_dataset.SALARY))]
full_opm_dataset



full_opm_dataset.to_csv("full_opm_dataset.csv")
