==> code/qss20_groupcode/dol_staffing/01_final_whd_cleaning_prepformerge.ipynb <==
import pandas as pd
import os
import itertools
from itertools import product
import numpy as np
import re 

os.getcwd()

##Drop all irrelevant columns in whd data
def column_clean(df):
    to_keep = ['case_violtn_cnt', 'naics_code_description', 'findings_end_date',
           'findings_start_date', 'case_id', 'h2a_violtn_cnt', 'h2a_bw_atp_amt',
           'h2a_ee_atp_cnt','h2a_cmp_assd_amt', 'st_cd', "ld_dt", "naics_cd"]   
    to_remove = [column for column in df.columns if column not in to_keep]
    return(df.drop(columns = to_remove))

whd = pd.read_csv("WHD data.csv")

# start of whd cleaning
whd = column_clean(whd)
##Subset to non-zero h2a violations 
whd = whd[~whd.h2a_violtn_cnt.astype(str).str.contains('0')]

whd = whd[~whd.ld_dt.astype(str).str.contains('1\d{3}', regex = True)]
whd = whd[~whd.ld_dt.astype(str).str.contains('200[0-9]', regex = True)]
whd = whd[~whd.ld_dt.astype(str).str.contains('201[0-1]', regex = True)]
whd = whd[~whd.ld_dt.astype(str).str.contains('2021', regex = True)]

# Create year and quarter columns in WHD data
whd["start_year"] = [int(date_string[-4:]) for date_string in whd.findings_start_date]
whd["end_year"] = [int(date_string[-4:]) for date_string in whd.findings_end_date]

regex = r"\d{1,2}"
whd["start_month_int"] = [int(re.match(regex, date_string).group()) for date_string in whd.findings_start_date]
whd["end_month_int"] = [int(re.match(regex, date_string).group()) for date_string in whd.findings_end_date]

month_to_qtr_dict = {1:"MAR", 2:"MAR", 3:"MAR", 4:"JUN", 5:"JUN", 6:"JUN", 7:"SEP", 8:"SEP", 9:"SEP", 10:"DEC", 11:"DEC", 12:"DEC"}
whd["start_quarter"] = [month_to_qtr_dict[month] for month in whd.start_month_int]
whd["end_quarter"] = [month_to_qtr_dict[month] for month in whd.end_month_int]


whd["ld_dt_object"] = pd.to_datetime(whd["ld_dt"])
whd["ld_dt_year"] = whd["ld_dt_object"].dt.year.astype(int)
whd["ld_dt_month"] = whd["ld_dt_object"].dt.month.astype(int)
whd["ld_dt_quarter"] = [month_to_qtr_dict[month] for month in whd.ld_dt_month]

whd.to_csv("whd_for_merge")

whd
