==> code/qss20_groupcode/dol_staffing/02_final_opm_wrangling_and_merge.ipynb <==
import pandas as pd
import os
import itertools
from itertools import product
import numpy as np
import re 

os.getcwd()

full_opm_dataset = pd.read_csv("full_opm_dataset.csv")
full_opm_dataset

whd = pd.read_csv("whd_for_merge")
whd

## recoding states to two-letter abbreviations
locdf = pd.read_csv("DTloc.txt")
locdf

loc_dict = {}
regex = r'[A-Z ]+'
for i in range(len(locdf.LOC)):
    loc_dict[locdf.LOC.iloc[i]] = re.findall(regex, locdf.LOCT.iloc[i])[0]

## code to load state crosswalk
## source- https://towardsdatascience.com/state-name-to-state-abbreviation-crosswalks-6936250976c
cw_location = 'http://app02.clerk.org/menu/ccis/Help/CCIS%20Codes/'
cw_filename = 'state_codes.html'
states = pd.read_html(cw_location + cw_filename)[0]
states["Description"].iloc[15] = "District of Columbia"

full_opm_dataset["STATE_NAME"] = [loc_dict[loc] for loc in full_opm_dataset.LOC] 

incodebook = []
for i in range(len(full_opm_dataset.STATE_NAME)):
    if full_opm_dataset.STATE_NAME.iloc[i] in list(states.Description.str.upper()):
       # print("here")
        incodebook.append(True)
    else:
        incodebook.append(False)
full_opm_dataset["incodebook"] = incodebook
full_opm_dataset = full_opm_dataset[full_opm_dataset.incodebook]


state_abbrev = []
for state_name in full_opm_dataset["STATE_NAME"]:
    for i in range(len(states.Description)):  
        if states.Description.iloc[i].upper() ==state_name:            
            state_abbrev.append(states.Code.iloc[i])
            
    
full_opm_dataset["STATE_ABBREV"] = state_abbrev

all_years = np.arange(2011, 2021)
all_quarters = ['MAR', 'JUN', 'SEP', 'DEC']
all_states = list(full_opm_dataset.STATE_ABBREV.value_counts().index)
all_states.append("PR")
total_combos = len(all_years) * len(all_quarters) * len(all_states)
all_years_repeated = list(np.repeat(all_years, len(all_quarters)*len(all_states)))
all_quarters_repeated = all_quarters*(len(all_states) * len(all_years))
all_year_quarters = pd.DataFrame({'YEAR': all_years_repeated,
                               'QUARTER': all_quarters_repeated})
all_state_years = pd.DataFrame({'STATE_ABBREV': list(np.repeat(all_states, len(list(all_years)))),
                                'YEAR': list(all_years) * len(all_states)})
all_quart_state = pd.merge(all_year_quarters,
                          all_state_years,
                          on = ['YEAR']).drop_duplicates()

all_quart_state

opm_year_qtr = full_opm_dataset.groupby(["YEAR", "QUARTER", "STATE_ABBREV"])
employment_totals = pd.DataFrame(opm_year_qtr["EMPLOYMENT"].sum())
average_salary = pd.DataFrame(opm_year_qtr["SALARY"].mean())
average_LOS = pd.DataFrame(opm_year_qtr["LOS"].mean())
opm_to_merge = pd.concat([employment_totals, average_salary, average_LOS], axis=1)


opm_to_merge

merged_opm_intermediate = pd.merge(all_quart_state, opm_to_merge, on=["YEAR", "QUARTER", "STATE_ABBREV"], how="left")
merged_opm_intermediate["EMPLOYMENT"] = merged_opm_intermediate["EMPLOYMENT"].fillna(0)
merged_opm = merged_opm_intermediate.set_index(["YEAR", "QUARTER", "STATE_ABBREV"], drop=True, append=False, inplace=False, verify_integrity=False)
merged_opm[merged_opm.EMPLOYMENT.isna()]

# We are going to assume that when combining the employment data with the violations data, the employment stats in the 4 quarters
# prior to ld_dt is what would have contributed to the finding of the violation since there is probably a lag between when 
# an investigation begins and when it is recorded.
# For example, for a violation with a loading year and quarter of DEC 2018, the total employment variable is the mean of the 
# number of emloyers for the quarters DEC 2017, MAR 2018, JUN 2018, and SEP 2018.

total_employment = []
for i in range(len(whd.ld_dt)):
    start_year = whd.ld_dt_year.iloc[i]
    start_quarter = whd.ld_dt_quarter.iloc[i]
    state = whd.st_cd.iloc[i]
    
    if start_quarter == "DEC":
        employment_4 = merged_opm.loc[((start_year - 1), "DEC", state),"EMPLOYMENT"]
        employment_3 = merged_opm.loc[(start_year, "MAR", state), "EMPLOYMENT"]
        employment_2 = merged_opm.loc[(start_year, "JUN", state), "EMPLOYMENT"]
        employment_1 = merged_opm.loc[(start_year, "SEP", state), "EMPLOYMENT"]
    elif start_quarter == "SEP":
        employment_4 = merged_opm.loc[((start_year - 1), "SEP", state),"EMPLOYMENT"]
        employment_3 = merged_opm.loc[((start_year-1), "DEC", state), "EMPLOYMENT"]
        employment_2 = merged_opm.loc[(start_year, "MAR", state), "EMPLOYMENT"]
        employment_1 = merged_opm.loc[(start_year, "JUN", state), "EMPLOYMENT"]
    elif start_quarter == "JUN":
        employment_4 = merged_opm.loc[((start_year - 1), "JUN", state),"EMPLOYMENT"]
        employment_3 = merged_opm.loc[((start_year-1), "SEP", state), "EMPLOYMENT"]
        employment_2 = merged_opm.loc[((start_year-1), "DEC", state), "EMPLOYMENT"]
        employment_1 = merged_opm.loc[(start_year, "MAR", state), "EMPLOYMENT"]
    elif start_quarter == "MAR":
        employment_4 = merged_opm.loc[((start_year - 1), "MAR", state),"EMPLOYMENT"]
        employment_3 = merged_opm.loc[((start_year-1), "JUN", state), "EMPLOYMENT"]
        employment_2 = merged_opm.loc[((start_year-1), "SEP", state), "EMPLOYMENT"]
        employment_1 = merged_opm.loc[(start_year, "DEC", state), "EMPLOYMENT"]
    
    mean_employees = (employment_4 + employment_3 + employment_2 + employment_1)/4 

    total_employment.append(round(mean_employees))

whd["total_employment"] = total_employment

final_dataset = whd.copy()
final_dataset.to_csv("final_dataset.csv")
final_dataset
