==> code/qss20_groupcode/dol_staffing/04_final_regression.ipynb <==
import pandas as pd
import os
import itertools
from itertools import product
import numpy as np
import re 

os.getcwd()

bw_data = pd.read_csv("rbw_csv.csv")
cmp_data = pd.read_csv("rcmp_csv.csv")

bw_data = bw_data.iloc[1:,:]
bw_data

cmp_data = cmp_data.iloc[1:,:]
cmp_data

bw_regression_data = bw_data.set_index(["st_cd", "ld_dt_year"])
bw_regression_data["backwages"] = [float(bw_regression_data.h2a_bw_atp_amt.iloc[i])/float(bw_regression_data.h2a_violtn_cnt.iloc[i]) for i in range(len(bw_regression_data.index))]
bw_regression_data["total_employment"] = [int(bw_regression_data.total_employment.iloc[i]) for i in range(len(bw_regression_data.total_employment))]
bw_regression_data

cmp_regression_data = cmp_data.set_index(["st_cd", "ld_dt_year"])
cmp_regression_data["cmp_amount"] = [float(cmp_regression_data.h2a_cmp_assd_amt.iloc[i])/float(cmp_regression_data.h2a_violtn_cnt.iloc[i]) for i in range(len(cmp_regression_data.index))]
cmp_regression_data["total_employment"] = [int(cmp_regression_data.total_employment.iloc[i]) for i in range(len(cmp_regression_data.total_employment))]
cmp_regression_data

from linearmodels import PanelOLS
mod1 = PanelOLS(bw_regression_data.backwages, bw_regression_data.total_employment, time_effects=True, entity_effects=True)
mod1fit = mod1.fit()
print(mod1fit)
print(mod1fit.summary.as_latex())

mod2 = PanelOLS(cmp_regression_data.cmp_amount, cmp_regression_data.total_employment, time_effects=True, entity_effects=True)
mod2fit = mod2.fit()
print(mod2fit)
print(mod1fit.summary.as_latex())
